checking owner added

+++ aiohttp/main_app.py
@@ -1,5 +1,6 @@
+import aiohttp.web
 from aiohttp import web
-from sqlalchemy.exc import IntegrityError
+from sqlalchemy.exc import IntegrityError, DBAPIError
 from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
 from sqlalchemy.future import select
 from sqlalchemy.orm import sessionmaker
@@ -9,7 +10,8 @@ from db import Post, Base, User
 from utils import hash_pw, check_pw
 
 # TODO env for everything
-PG_DSN = 'postgresql+asyncpg://app:1234@db:5432/test_db'
+# PG_DSN = 'postgresql+asyncpg://app:1234@db:5432/test_db'
+PG_DSN = 'postgresql+asyncpg://app:1234@localhost:5431/test_db'
 
 engine = create_async_engine(PG_DSN)
 Session = sessionmaker(bind=engine, expire_on_commit=False, class_=AsyncSession)
@@ -24,13 +26,41 @@ async def orm_context(app: web.Application):
 
 
 @web.middleware
-async def session_middleware(requests: web.Request, handler):
+async def session_middleware(request: web.Request, handler):
     async with Session() as session:
-        requests['session'] = session
-        return await handler(requests)
+        request['session'] = session
+        return await handler(request)
 
 
+@web.middleware
+async def authentication_middleware(request: web.Request, handler):
+    req_token = request.headers.get('token')
+    if not req_token:
+        raise web.HTTPForbidden(
+            text=json.dumps({'status': 'error', 'message': 'enter the token'}),
+            content_type='application/json',
+        )
 
+    session = request['session']
+    db_query = select(User).where(User.token == req_token)
+    try:
+        db_res = await session.execute(db_query)
+    except DBAPIError:
+        raise web.HTTPForbidden(
+            text=json.dumps({'status': 'error', 'message': 'enter correct length token'}),
+            content_type='application/json',
+        )
+    user = db_res.scalar()
+    request['token'] = [user.token, user.id]
+    return await handler(request)
+
+
+def check_owner(request, user_id):
+    if not request['token'] or request['token'][1] != user_id:
+        raise web.HTTPForbidden(
+            text=json.dumps({'status': 'error', 'message': 'only owner has access'}),
+            content_type='application/json',
+        )
 async def get_post(post_id: int, session: Session):
     post = await session.get(Post, post_id)
 
@@ -41,6 +71,7 @@ async def get_post(post_id: int, session: Session):
         )
 
     return post
+
 
 async def get_user(user_id: int, session: Session):
     user = await session.get(User, user_id)
@@ -107,13 +138,13 @@ class UserView(web.View):
 
     async def get(self):
         user_id = int(self.request.match_info['user_id'])
+        check_owner(self.request, user_id)
         session = self.request['session']
         user = await get_user(user_id, session)
         return web.json_response({
             'id': user.id,
             'name': user.name,
             'email': user.email,
-            'token': str(user.token),
         })
 
     async def post(self):
@@ -134,6 +165,7 @@ class UserView(web.View):
     async def patch(self):
         session = self.request['session']
         user_id = int(self.request.match_info['user_id'])
+        check_owner(self.request, user_id)
         user_data = await self.request.json()
         if 'password' in user_data:
             user_data['password'] = hash_pw(user_data['password'])
@@ -147,15 +179,17 @@ class UserView(web.View):
     async def delete(self):
         session = self.request['session']
         user_id = int(self.request.match_info['user_id'])
+        check_owner(self.request, user_id)
         user = await get_user(user_id, session)
         await session.delete(user)
         await session.commit()
         return web.json_response({'status': 'delete success'})
 
+
 async def login(request):
     session = request['session']
     json_data = await request.json()
-    db_query = select(User).where(User.name==json_data['name'])
+    db_query = select(User).where(User.name == json_data['name'])
     db_res = await session.execute(db_query)
     user = db_res.scalar()
     if not user or not check_pw(json_data['password'], user.password):
@@ -170,6 +204,7 @@ async def main():
     app = web.Application()
     app.cleanup_ctx.append(orm_context)
     app.middlewares.append(session_middleware)
+    app.middlewares.append(authentication_middleware)
     app.add_routes([
         web.get('/posts/{post_id:\d+}/', PostView),
         web.post('/posts/', PostView),
@@ -185,3 +220,6 @@ async def main():
         web.post('/login/', login)
     ])
     return app
+
+app = main()
+aiohttp.web.run_app(app, port=8081)
